FontAccess: Replace Sticky Activation with Page Visibility check
This CL replaces the Sticky Activation check for enumeration with a page
visibility check. The attack vector that the sticky activation check was
intended to mitigate is background access to the API.

However, this prevents an intended use-case of the API: e.g. navigating
to a link to open a document will require user activation before content
is show if this API is necessary to render on-screen.

Using a Page Visibility change instead mitigates the same attack vector
while enabling the navigation use-case.

A manual test will be added in a subsequent CL.

Bug: 1043306
Change-Id: Ieb12341f78bceec94907fb577f3c969ae839abdd